remove add add friends logic added, Model corrected

+++ controllers/userController.js
@@ -13,7 +13,12 @@ exports.getAllUsers = async (req, res) => {
 exports.getUser = async (req, res) => {
   try {
     const { userId } = req.params;
-    const user = await User.findById(userId);
+    if (!userId) {
+      return res.status(406).json('userId must be provided');
+    }
+    const user = await User.findById(userId)    
+    .populate("thoughts")
+    .populate("friends");
 
     res.status(200).json(user);
   } catch (err) {
@@ -24,9 +29,12 @@ exports.getUser = async (req, res) => {
 exports.createUser = async (req, res) => {
   try {
     const { username, email } = req.body;
+    if (!username || !email) {
+      return res.status(406).json('username and email must be provided');
+    }
     const user = await User.create({ username, email });
-      // .populate("thoughts")
-      // .populate("friends");
+    // .populate("thoughts")
+    // .populate("friends");
 
     res.status(201).json(user);
   } catch (err) {
@@ -38,6 +46,11 @@ exports.updateUser = async (req, res) => {
   try {
     const { userId } = req.params;
     const { username, email } = req.body;
+    
+    if (!userId || !username || !email) {
+      return res.status(406).json('userId, username and email must be provided');
+    }
+
     const updatedUser = await User.findOneAndUpdate(
       { id: userId },
       {
@@ -58,10 +71,51 @@ exports.updateUser = async (req, res) => {
 exports.deleteUser = async (req, res) => {
   try {
     const { userId } = req.params;
-  
+    if (!userId) {
+      return res.status(406).json('userId must be provided');
+    }
+
     const deletedUser = await User.deleteOne({ id: userId });
 
     res.status(204).json(deletedUser);
+  } catch (err) {
+    return res.status(500).json(err);
+  }
+};
+
+exports.addFriendToUsersFriendList = async (req, res) => {
+  try {
+    const { userId, friendId } = req.params;
+
+    if (!userId || !friendId) {
+      return res.status(406).json('userId and friendId must be provided');
+    }
+    const user = await User.findOneAndUpdate(
+      { id: userId },
+      { $addToSet: { friends: friendId } },
+      { new: true }
+    );
+
+    res.status(201).json(user);
+  } catch (err) {
+    return res.status(500).json(err);
+  }
+};
+
+exports.removeFriendFromUsersFriendList = async (req, res) => {
+  try {
+    const { userId, friendId } = req.params;
+
+    if (!userId || !friendId) {
+      return res.status(406).json('userId and friendId must be provided');
+    }
+    const user = await User.findOneAndUpdate(
+      { id: userId },
+      { $pull: { friends: friendId } },
+      { new: true }
+    );
+
+    res.status(201).json(user);
   } catch (err) {
     return res.status(500).json(err);
   }
